+++ tests/DependencyInjection/Fixtures/php/full.php
@@ -17,6 +17,7 @@ $container->loadFromExtension('sentry', [
         'sample_rate' => 1,
         'traces_sample_rate' => 1,
         'traces_sampler' => 'App\\Sentry\\Tracing\\TracesSampler',
+        'trace_propagation_targets' => ['website.invalid'],
         'attach_stacktrace' => true,
         'context_lines' => 0,
         'enable_compression' => true,