+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat: Add support for tracing of the Symfony HTTP client requests (#606)
 - feat: Support logging the impersonator user, if any (#647)
 - ref: Use constant for the SDK version (#662)
+- Add support for HTTP client baggage propagation (#663)
 
 ## 4.4.0 (2022-10-20)